consolidate logic between the reducers
